Transferred GameServiceImpl methods to MockDAO class. Added update and delete methods.
